requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.7
decorator==5.1.1
dictdiffer==0.9.0
diffusers==0.25.0
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
dpath==2.1.6
dulwich==0.21.7
dvc==3.38.1
dvc-data==3.7.0
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==3.0.5
dvc-render==1.0.0
dvc-studio-client==0.18.0
dvc-task==0.3.0
entrypoints==0.4
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gcsfs==2023.12.2.post1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grandalf==0.8
gto==1.6.2
httplib2==0.22.0
huggingface-hub==0.20.2
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
iterative-telemetry==0.0.8
Jinja2==3.1.2
kombu==5.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.9.10
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
pipreqs==0.4.13
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scmrepo==2.0.2
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sqltrie==0.11.0
sympy==1.12
tabulate==0.9.0
tokenizers==0.15.0
tomlkit==0.12.3
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
voluptuous==0.14.1
wcwidth==0.2.13
yarg==0.1.9
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
datasets==2.16.1
joblib==1.3.2
scikit-learn==1.3.2
scipy==1.11.4
